Support CryptoNight Variant 1 Algorithm
- Use bitbandi's node-multi-hashing instead of Snipa22's
- Update lib/coins/xmr.js to support switching to CN V1 on v7 of the blockchain
- Update lib/coins/{itns,itns142,krb}.js to hardcode CN V0 until such time that they
  decide (if they decide) to fork to support CN V1.  We'll just need to know the
  version of their chain that they plan to do this for in order to patch support in.

While I would have loved to keep Snipa22's node-multi-hashing-aesni in place,
I do not have the knowledge to patch it to support the new CN Variant. So, this
commit modifies multi-hashing to be pulled from bitbandi's repo.

If Snipa22's repo gets patched by someone who understands those changes, I will
look into reverting back to it.

Tracking bitbandi's repo is a bit of an unknown at this time.  They seem to be active
on github and that's about it.  But they are tracking zone117x's repo, and moneromooo
of XMR fame submit the CN V1 patch to his repo. We are using bitbandi's because they
patched in support for newer node versions, which we use for nodejs-pool.